==> src/01_sklearn/00_sampling_importance.ipynb <==
import pandas as pd
import networkx as nx
import numpy as np
import math
import random
import csv

train_csv = pd.read_csv('./data/individual_links.txt', names=['source', 'destination'],sep='\s+')
g = nx.from_pandas_edgelist(train_csv, source='source',target='destination',create_using=nx.DiGraph())    

num_edges = g.number_of_edges()
num_nodes = g.number_of_nodes()
print(num_edges)
print(num_nodes)

node_degree_dict=nx.degree_centrality(g)

sorted_node_degree_list = sorted(node_degree_dict.items(), key=lambda x: x[1], reverse=True)    

sorted_node_degree_list

# Get top Million
N = 1000000
sample_sorted_node_degree_list = sorted_node_degree_list[0:N]

#randomly select 10% edges as training edges
#num_test_edges = int(math.floor(0.01 * num_edges))

#num_test_edges = 20000
num_test_edges = 50000



#edges_pos = random.sample(g.edges(), num_test_edges)

i = 0
edges_pos = []
while i < num_test_edges:
    node = random.sample(sample_sorted_node_degree_list, 1)
    node = node[0][0]
    try:
        # get neighbors from node
        neighbors = list(g.neighbors(node))
        random_neighbor = random.sample(neighbors,1)[0]
        
        edges_pos.append([node,random_neighbor])
        i = i+1
    except Exception as e:
        #print(e)
        pass

with open("edges_pos_importance_sampling.csv","w") as csvfile:
    writer=csv.writer(csvfile)
    writer.writerow(["Source", "Target", "Label"])
    for edge in edges_pos:
        writer.writerow([edge[0], edge[1],1])

i = 0
edges_neg = []
while i < num_test_edges:
    edge = random.sample(sample_sorted_node_degree_list, 2)
    source = edge[0][0]
    target = edge[1][0]

    try:
        edge_exists = g.has_edge(source,target)
        if edge_exists == False:
            edges_neg.append([source, target])
            i = i+1
    except Exception as e:
        print(e)
        pass

with open("edges_neg_importance_sampling.csv","w",newline="") as csvfile:
    writer=csv.writer(csvfile)
    writer.writerow(["Source","Target", "Label"])
    for edge in edges_neg:
        writer.writerow([edge[0], edge[1], 0])
