==> src/01_sklearn/02_LogisticRegression.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

# Importing the dataset
dataset = pd.read_csv('train.csv')
FEATURE_SIZE=3

X = dataset.iloc[:,0:FEATURE_SIZE].values
y = dataset.iloc[:, FEATURE_SIZE].values

# Splitting the dataset into the Training set and Test set
from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 0)
# Feature Scaling
from sklearn.preprocessing import StandardScaler
sc = StandardScaler()
x_train = sc.fit_transform(x_train)
x_test = sc.transform(x_test)

from sklearn import linear_model
lr=linear_model.LogisticRegression(C=1e4,penalty='l2')
lr.fit(x_train,y_train)

from sklearn.metrics import roc_auc_score
pre=lr.predict_proba(x_test)
# print(pre)
y_pre=[p[1] for p in pre]
acc=lr.score(x_test,y_test)
print(acc)
auc=roc_auc_score(y_test,y_pre)
print(auc)


y

testdata = pd.read_csv('test.csv')
x_testing = testdata.iloc[:,0:FEATURE_SIZE].values
x_testing = sc.transform(x_testing)
predictions=lr.predict_proba(x_testing)

import csv
with open("kaggle.csv","w",newline="") as csvfile:
    writer=csv.writer(csvfile)
    writer.writerow(["Id","Predicted"])
    test_id=1
    for prediction in predictions:
        writer.writerow([test_id,prediction[1]])
        test_id+=1
