==> src/01_sklearn/04_node2vec.ipynb <==
import gc
import networkx as nx
import numpy as np
import os
import pandas as pd
import time
import scipy
import sklearn
from sklearn import cluster, linear_model
from sklearn.decomposition import TruncatedSVD
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.model_selection import train_test_split
from sklearn.multiclass import OneVsRestClassifier
import sys
import warnings # Silence perf warning

sys.path.append(os.path.realpath('..'))

import nodevectors
import csrgraph as cg
from csrgraph import methods
from nodevectors.evaluation import link_pred


import csrgraph as cg
import nodevectors

ggvec_params = dict(
    order=2,
    tol=0.07,
    tol_samples=10,
    max_epoch=6_000,
    learning_rate=0.1,
    negative_ratio=0.15,
    exponent=0.33,
    verbose=True,
)

train_csv = pd.read_csv('./data/output_1.csv', names=['source', 'destination'],sep='\s+')
G = nx.from_pandas_edgelist(train_csv, source='source',target='destination')
    

G = cg.read_edgelist('./data/output_1.csv', sep='\s+')

w_train = nodevectors.GGVec(**ggvec_params).fit_transform(G)
